auth_project/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
google-auth
google-auth-oauthlib 
google-auth-httplib2
requests
